requirements.txt
python-telegram-bot==13.7
APScheduler==3.6.3